Refactored common pieces out into HttpObject.
